Automatic Instagram direct message
This is a Python program to login into your Instagram id and follow/send message to a particular profile. you have to enter - your username, your password , the receiver username and the message. this is a program to send message to single person per run but can be easily modified to send same message to multiple people at a single run.
PS- inside the code, if you just want to follow - remove the message part from code and if you want to send message, you have to be following the person. if you already follow remove follow part from the code